Move comparison to progress VC
The photo report VC is now just responsible for creating a
FingerPrint for the progress view to use in the comparison.
It passes this to the progress view via the segue.

In theory, the progress bar should now update in real time,
as each comparison completes. In practice it jumps to around
10%, pauses and then jumps to 100%. This is on the simulator;
perhaps it will behave differently on a Real Device.

Next we'll close the progress view once all comparisons have
run, as well as pass along the results to the next VC to
display.